requirements.txt
typer
shellingham
python-dotenv
loguru
pydantic
pydantic-settings
icecream
redis
litellm
nbformat
ansi2html
e2b_code_interpreter
jupyter_client
fastapi
uvicorn
pypandoc
